clear next.tsx
